### 2019-11-25 #1916 @Kevin-Mattheus-Moerman
@Sulstice your paper seems to be missing some fields (e.g. affiliations).

Have a look here:
https://joss.readthedocs.io/en/latest/submitting.html#checking-that-your-paper-compiles

You can test compile your paper here:
https://whedon.theoj.org/
