### 2019-11-25 19b7d00
ss/MASTER added a minor tweak to the paper
