### 2019-11-26 c3fb311
ss/MASTER added addendums to the paper.md
