### 2019-11-26 #1916 @Sulstice
Hi @Kevin-Mattheus-Moerman I just updated the affiliations to "No affiliations" since I currently have none. Does that suffice or is there another protocol I should follow?
