### 2019-11-26 #1916 @arfon
@Sulstice - you can add an affiliation of 'None' but this need to be in the `affiliations` block as described in the example paper: https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography
