### 2019-11-26 8477ae0
ss/MASTER added the include proline functions, supressed RDKit output and nixed the load_datasources function and directory
