+++ cocktail_shaker/peptide_builder.py
@@ -17,8 +17,9 @@ class PeptideMolecule(object):
     __version__ = '1.1.0'
     __update__ = 'False'
 
-    def __init__(self, length_of_peptide = 1):
+    def __init__(self, length_of_peptide = 1, include_proline = False):
         self.length_of_peptide = int(length_of_peptide)
+        self.include_proline = include_proline
         self.peptide_base = self.build_peptide_backbone()
         self.peptide_replaced = self.build_peptide_molecule_temp_replacements()
 
@@ -49,6 +50,10 @@ class PeptideMolecule(object):
             peptide_c_terminus += '=O)'
 
         peptide_backbone = peptide_n_terminus + peptide_c_terminus[:-1]
+
+        if self.include_proline:
+            peptide_backbone = self._add_proline(peptide_backbone)
+
         self.peptide_base = peptide_backbone
 
         return peptide_backbone
@@ -66,6 +71,23 @@ class PeptideMolecule(object):
             replaced_peptides = replaced_peptides[0:start_index_pattern + 1] + 'C([*:' + str(i) + '])' + replaced_peptides[start_index_pattern + 2:]
 
         return replaced_peptides
+
+    def _add_proline(self, peptide_backbone):
+
+        """
+
+        Adds the proline on the N-Terminus
+
+        Arguments:
+            peptide_backbone (String): Peptide Backbone Chain without slots
+
+        Returns:
+            proline_peptide_backbone (String): Modified peptide_backbone with the proline addition into the n-terminus.
+        """
+        proline_peptide_backbone = peptide_backbone.replace("NC", "N2CCCC2", 1)
+
+        return proline_peptide_backbone
+
 
     def build_c_terminus(self):
 
@@ -86,3 +108,4 @@ class PeptideMolecule(object):
         """
 
         pass
+
