+++ cocktail_shaker/datasources/shapes.yaml
@@ -1,13 +1,0 @@
-Shapes:
-
-  # Rods
-  # ----
-  Rods: []
-
-  # Spheres
-  # -------
-  Spheres: []
-
-  # Discs
-  # -----
-  Discs: []

[…]
