### 2019-11-26 88c4761
ss/MASTER updated some irregularities in the docs
