### 2019-11-27 79c6154
Update pygments from 2.4.2 to 2.5.1
