+++ requirements.txt
@@ -20,7 +20,7 @@ pluggy==0.13.1
 progressbar==2.5
 progressbar2==3.47.0
 py==1.8.0
-Pygments==2.4.2
+Pygments==2.5.1
 pyparsing==2.4.5
 pytest==5.3.0
 python-dateutil==2.8.1
