+++ requirements.txt
@@ -22,7 +22,7 @@ progressbar2==3.47.0
 py==1.8.0
 Pygments==2.4.2
 pyparsing==2.4.5
-pytest==5.3.0
+pytest==5.3.1
 python-dateutil==2.8.1
 python-utils==2.3.0
 pytz==2019.3
