### 2019-11-29 8172336
Update pygments from 2.4.2 to 2.5.2
