### 2019-12-01 42c3413
Update paper.md
