### 2019-12-01 f1789a3
Update paper.md
