### 2019-12-01 0c94371
Update importlib-metadata from 0.23 to 1.1.0
