### 2019-12-02 #1916 @csoneson
Hi @Sulstice - could you check Figure 1 in your paper? It doesn't seem to show up.
