### 2019-12-02 #1916 @arfon
@Sulstice - you'll need to link to the image URL like this rather than the relative path you're currently using:

```
https://raw.githubusercontent.com/Sulstice/cocktail-shaker/master/images/figure_1.png
```
