+++ paper.md
@@ -42,7 +42,7 @@ of whether to include generation of stereoisomers and/or natural amino acids.
 Cocktail Shaker will generate all combinations of the library and allocate them to a slot within the peptide. This process of 
 string manipulation is presented in ```Figure 1```.
 
-![Full string manipulation diagram of how cocktail shaker works with a ligand library of just bromine and iodine. 1D representations are labeled above with their 2D depictions displayed below.](./images/figure_1.png)
+![Full string manipulation diagram of how cocktail shaker works with a ligand library of just bromine and iodine. 1D representations are labeled above with their 2D depictions displayed below.](https://raw.githubusercontent.com/Sulstice/cocktail-shaker/master/images/figure_1.png)
   
 
 ```Cocktail Shaker``` also allows for File Writing of the molecules into a wide array of chemical formats (found in the documentation).
