### 2019-12-03 c774707
Update sphinx from 2.2.1 to 2.2.2
