+++ paper.bib
@@ -1,11 +1,38 @@
-@article{cocktail-shaker,
-    author = {{Suliman Sharif}},
-    title = "{Cocktail Shaker: An open source drug expansion and enumeration library using for peptides}",
-    journal = {\aap},
-    keywords = {methods: cheminformatics, drug design, peptides, chemistry methods: python3, RDKit},
-    year = 2013,
-    month = nov,
-    volume = 558,
-    doi = {},
-    adsurl = {www.github.com/Sulstice/Cocktail-Shaker}
+@article{Craig:2013-1,
+    author = {{Craig}, D. and {Fairlie}, D. and {Liras} S. and {Price} D.},
+    title = "The future of peptide-based drugs.",
+    journal = "Chemical Biology & Drug Design",
+    year = "2013",
+    doi = "10.1111/cbdd.12055"
 }
+
+@article{Bruno:2013-2,
+    author = {{Bruno}, B. and {Miller}, G. and {Lim} C.},
+    title = "Basics and recent advances in peptide and protein drug delivery",
+    journal = "Therapeutic Delivery",
+    year = "2013",
+    doi = "10.4155/tde.13.104"
+}
+
+@article{Reynolds:2010-3,
+    author = {{Reynolds}, C. and {Merz}, K. and {Ringe} D.},
+    title = "Drug Design: Structure- and Ligand-Based Approaches",
+    journal = "Cambrige University Press",
+    year = "2010",
+    ISBN = "978-0521887236"
+}
+
+@article{Maillet:2020-4,
+    author = {{Malliet}, N.},
+    title = "Generator: fast and efficient in silico protein digestion",
+    journal = "NAR Genomics and Bioinformatics",
+    year = "2020",
+    doi = "10.1093/nargab/lqz004"
+}
+
+@article{Landrum:2019-5,
+    author = {{Landrum}, G.},
+    title = "RDKit: Open-Source Cheminformatics Software",
+    year = "2019",
+}
+
