### 2019-12-05 #1916 @Sulstice
@csoneson I looked through the list

Username: mskwark, road2stat, semacu seem most reasonable since they are in the peptide/cheminformatics space.
