### 2019-12-05 #1916 @Sulstice
Should I be doing a release of the software so y'all can pip install it? Or should I wait pending on whether you approve or not?
