### 2019-12-05 9ba214e
Update importlib-metadata from 0.23 to 1.2.0
