### 2019-12-06 #1916 @csoneson
@Sulstice - yes, I think that would be nice.
