### 2019-12-06 #1916 @csoneson
@mskwark, @semacu - would one of you be willing to review this submission for JOSS?
