### 2019-12-06 3ad7d21
ss/MASTER changed versions of the classes
