### 2019-12-06 d2694f2
ss/MASTER made some quick changes
