+++ cocktail_shaker/functional_group_enumerator.py
@@ -20,7 +20,7 @@ import progressbar
 
 # Cocktail Shaker Imports
 # -----------------------
-from validation import MoleculeValidator
+from .validation import MoleculeValidator
 
 class Cocktail(object):
     """
