+++ cocktail_shaker/file_handler.py
@@ -9,7 +9,7 @@
 from rdkit import Chem
 from pathlib import Path
 import progressbar
-from request_handler import CactusRequestHandler, Resolver
+from .request_handler import CactusRequestHandler, Resolver
 
 class FileNotSupportedError(Exception):
 
