### 2019-12-07 #1916 @Sulstice
@csoneson I released the cocktail shaker v1.1.0 and it is labeled as a "pre-release" for now pending the JOSS approval.

You should be able to easily install with this command `pip install cocktail-shaker==1.1.0`
