### 2019-12-10 f473db8
ss/MASTER added in a time sleepeer to handle the NIH resolver dropping requests
