### 2019-12-10 ec8c31f
ss/MASTER added a cert requirement
