### 2019-12-10 aaf97fc
ss/MASTER added the function without a timeout
