### 2019-12-11 #1916 @csoneson
Ping @mskwark, @semacu

> @mskwark, @semacu - would one of you be willing to review this submission for JOSS?
