### 2019-12-12 3205f1b
Update importlib-metadata from 0.23 to 1.3.0
