### 2019-12-14 #1916 @csoneson
@olivertomic, @road2stat - would one of you be willing to review this submission for JOSS?
