### 2019-12-15 bca6a9b
Update pytest from 5.3.0 to 5.3.2
