### 2019-12-16 #1916 @csoneson
@diegoenry, @matthiaslein - would one of you be willing to review this submission for JOSS?
