### 2019-12-19 #1916 @csoneson
@fdroessler, @keipertk - would one of you be interested in reviewing this submission for JOSS?
