### 2019-12-23 a698dea
Update numpy from 1.17.3 to 1.18.0
