### 2019-12-26 8e87981
Update imagesize from 1.1.0 to 1.2.0
