+++ requirements.txt
@@ -4,7 +4,7 @@ Babel==2.7.0
 certifi==2019.9.11
 chardet==3.0.4
 idna==2.8
-imagesize==1.1.0
+imagesize==1.2.0
 importlib-metadata==0.23
 Jinja2==2.10.3
 lxml==4.4.1
