+++ requirements.txt
@@ -19,7 +19,7 @@ Pillow==6.2.1
 pluggy==0.13.1
 progressbar==2.5
 progressbar2==3.47.0
-py==1.8.0
+py==1.8.1
 Pygments==2.4.2
 pyparsing==2.4.5
 pytest==5.3.0
