+++ requirements.txt
@@ -1,6 +1,6 @@
 alabaster==0.7.12
 attrs==19.3.0
-Babel==2.7.0
+Babel==2.8.0
 certifi==2019.9.11
 chardet==3.0.4
 idna==2.8
