### 2020-01-02 ce7605e
Update wcwidth from 0.1.7 to 0.1.8
