### 2020-01-03 #1916 @csoneson
👋 @olivertomic - Happy New Year! Would you still be interested and available in mid-February to review this submission for JOSS?

> Cocktail Shaker: An open source drug expansion and enumeration library for peptides
