### 2020-01-04 #1992 @whedon
[REVIEW]: Cocktail Shaker: An open source drug expansion and enumeration library for peptides
**Submitting author:** @Sulstice (<a href="http://orcid.org/0000-0002-1342-9258">Suliman Sharif</a>)
**Repository:** <a href="https://github.com/Sulstice/cocktail-shaker" target ="_blank">https://github.com/Sulstice/cocktail-shaker</a>
**Version:** 1.1.8
