### 2020-01-07 b2a497b
Update numpy from 1.17.3 to 1.18.1
