+++ requirements.txt
@@ -11,7 +11,7 @@ lxml==4.4.1
 MarkupSafe==1.1.1
 MolVS==0.1.1
 more-itertools==7.2.0
-numpy==1.17.3
+numpy==1.18.1
 olefile==0.46
 packaging==19.2
 pandas==0.25.3
