### 2020-01-11 b5adbfc
Update importlib-metadata from 0.23 to 1.4.0
