+++ requirements.txt
@@ -5,7 +5,7 @@ certifi==2019.9.11
 chardet==3.0.4
 idna==2.8
 imagesize==1.1.0
-importlib-metadata==0.23
+importlib-metadata==1.4.0
 Jinja2==2.10.3
 lxml==4.4.1
 MarkupSafe==1.1.1
