### 2020-01-11 6bbd721
MASTER added the new functionality for cocktail shaker, circular peptides, pickling, and drug compound filtering
