+++ cocktail_shaker/functional_group_enumerator.py
@@ -17,6 +17,7 @@ lg.setLevel(RDLogger.CRITICAL)
 import ruamel.yaml as yaml
 import itertools
 import progressbar
+import functools
 
 # Cocktail Shaker Imports
 # -----------------------
@@ -54,7 +55,6 @@ class Cocktail(object):
         # I will reimplement it as a list.
 
         self.peptide_backbone = str(peptide_backbone)
-        print(peptide_backbone)
         self.ligand_library = ligand_library
 
         # Detect the proline amino acid on the N-terminus, set the max peptide length accordingly.
@@ -65,17 +65,27 @@ class Cocktail(object):
         self.enable_isomers = enable_isomers
         self.include_amino_acids = include_amino_acids
 
+        # Experimental cache argument.
+        self.cache = False
+
         self.combinations = []
 
         if (self.peptide_backbone_length > len(self.ligand_library) or not ligand_library) and self.include_amino_acids == False:
             print ("Cocktail Shaker Error: Peptide Backbone Length needs to be less than or equal to for your library")
             raise IndexError
 
-    def shake(self):
+    def shake(self, store_as_pickle = False, compound_filters = []):
 
         """
 
         Generate all combinations of a molecule
+
+        Arguments:
+            store_as_pickle (Bool): If the user wants to store the results into a pickle file.
+            compound_filters (List): List of filters that the user would like to apply to their library.
+
+        Returns:
+            results (List): List of compounds generated into SMILES format.
 
         """
 
@@ -101,15 +111,16 @@ class Cocktail(object):
                                                                Chem.MolFromSmiles('[*:'+ str(j+1)+']'),
                                                                Chem.MolFromSmiles(str(combination[j])))
 
-                peptide_molecule = Chem.MolToSmiles(modified_molecule[0], isomericSmiles=True)
+                peptide_molecule = Chem.MolToSmiles(modified_molecule[0], isomericSmiles=True, canonical=True)
 
             # Enable StereoChemistry
             if self.enable_isomers:
 
+                from rdkit.Chem import EnumerateStereoisomers
                 molecule = Chem.MolFromSmiles(str(peptide_molecule))
-                options = Chem.EnumerateStereoisomers.StereoEnumerationOptions(unique=True, tryEmbedding=True)
-                isomers = tuple(Chem.EnumerateStereoisomers.EnumerateStereoisomers(molecule, options=options))
-                for smile in sorted(Chem.MolToSmiles(isomer, isomericSmiles=True) for isomer in isomers):
+                options = EnumerateStereoisomers.StereoEnumerationOptions(unique=True, tryEmbedding=True)
+                isomers = tuple(EnumerateStereoisomers.EnumerateStereoisomers(molecule, options=options))
+                for smile in sorted(Chem.MolToSmiles(isomer, isomericSmiles=True,  canonical=True) for isomer in isomers):
                     results.append(str(smile))
             else:
                 results.append(str(peptide_molecule))
@@ -122,6 +133,18 @@ class Cocktail(object):
 
         # Store in combinations if enumeration
         self.combinations = results
+
+        # Handle storage into a pickle file.
+        if store_as_pickle:
+
+            import pickle
+            with open("cocktail_shaker.pickle", 'wb') as cache_handle:
+                print("saving result to cache '%s'" % "cocktail_shaker.pickle")
+                pickle.dump(results, cache_handle)
+
+        if compound_filters:
+            results = self._apply_drug_filters(compound_filters)
+            self.combinations = results
 
         return results
 
@@ -217,5 +240,141 @@ class Cocktail(object):
 
         return natural_amino_acids
 
-        
+    def _apply_drug_filters(self, filters):
 
+        """
+
[…]
