+++ cocktail_shaker/peptide_builder.py
@@ -4,7 +4,21 @@
 #
 # ----------------------------------------------------------
 
-class PeptideMolecule(object):
+class CircularPeptideError(Exception):
+
+    __version_error_parser__ = "1.1.0"
+    __allow_update__ = False
+
+    """
+
+    Raise File Not Supported Error if we don't support the parsing. 
+
+    """
+    def __init__(self, message, errors):
+        super().__init__(message)
+        self.errors = errors
+
+class PeptideBuilder(object):
 
     """
 
@@ -15,13 +29,24 @@ class PeptideMolecule(object):
     """
 
     __version__ = '1.1.0'
-    __update__ = 'False'
+    __update__ = False
 
-    def __init__(self, length_of_peptide = 1, include_proline = False):
+    # Decrease RAM time.
+    __slots__ = ['length_of_peptide', 'include_proline', 'circular', 'peptide_base', 'peptide_replaced']
+
+    def __init__(self, length_of_peptide = 1, include_proline = False, circular = False):
         self.length_of_peptide = int(length_of_peptide)
         self.include_proline = include_proline
-        self.peptide_base = self.build_peptide_backbone()
-        self.peptide_replaced = self.build_peptide_molecule_temp_replacements()
+        self.circular = circular
+
+        if self.circular:
+            if self.length_of_peptide <= 2:
+                print ("Circular peptides require a minimum of 3 amino acids in length")
+            self.peptide_base = self.build_circular_peptide_backbone()
+            self.peptide_replaced = self.build_peptide_molecule_temp_replacements()
+        else:
+            self.peptide_base = self.build_peptide_backbone()
+            self.peptide_replaced = self.build_peptide_molecule_temp_replacements()
 
     def __repr__(self):
         return str(self.peptide_replaced)
@@ -58,6 +83,36 @@ class PeptideMolecule(object):
 
         return peptide_backbone
 
+    def build_circular_peptide_backbone(self):
+
+        """
+
+        Builds the circular peptide backbone string
+
+        Returns:
+            peptide (String): Peptide molecule string of a certain length
+            peptide_base (String): The base peptide of 1 length amino acid.
+
+        """
+
+        # Return the base of the peptide for one length.
+        if self.length_of_peptide == 3:
+            return 'O=C1CNC(=O)CNC(=O)CN1'
+
+        circular_peptide_n_terminus = 'CN1'
+        circular_peptide_c_terminus = 'O=C1CNC(=O)'
+
+        pattern = ''
+
+        for i in range(self.length_of_peptide - 2):
+            pattern += 'CNC(=O)'
+
+        peptide_backbone = circular_peptide_c_terminus + pattern + circular_peptide_n_terminus
+
+        self.peptide_base = peptide_backbone
+
+        return peptide_backbone
+
     def build_peptide_molecule_temp_replacements(self):
 
         """
@@ -65,10 +120,25 @@ class PeptideMolecule(object):
         Build a temporary functional peptide molecule string to be used for the combinations downstream.
 
         """
+
         replaced_peptides = self.peptide_base
-        for i in range(1, self.length_of_peptide + 1):
-            start_index_pattern = replaced_peptides.find('NCC')
-            replaced_peptides = replaced_peptides[0:start_index_pattern + 1] + 'C([*:' + str(i) + '])' + replaced_peptides[start_index_pattern + 2:]
+
+        if self.circular:
+            for i in range(1, self.length_of_peptide + 1):
+
+                # Special case for circular peptides
+                # Last carbon that closes the ring is pattern = CN1
+                if i == self.length_of_peptide:
+                    start_index_pattern = replaced_peptides.find('CN1')
[…]

[…]
