### 2020-01-11 22006ff
ss/MASTER added a new test
