+++ tests/test_r_group_enumerator.py
@@ -180,23 +180,23 @@ def test_drug_filters():
         length_of_peptide = 4,
         circular = True,
     )
-
-    cocktail = Cocktail(peptide_backbone, ligand_library = ["Br", "Cl", "I", "F"])
-
-    combinations = cocktail.shake(compound_filters=["Lipinski"])
-    assert 'O=C1NC(Br)C(=O)NC(I)C(=O)NC(Cl)C(=O)NC1F' in combinations
-
-    combinations = cocktail.shake(compound_filters=["Ghose"])
-    assert len(combinations) == 0
-
-    combinations = cocktail.shake(compound_filters=["Veber"])
-    assert 'O=C1NC(I)C(=O)NC(Br)C(=O)NC1F' in combinations
-
-    combinations = cocktail.shake(compound_filters=["Rule of 3"])
-    assert len(combinations) == 0
-
-    combinations = cocktail.shake(compound_filters=["REOS"])
-    assert 'O=C1NC(Br)C(=O)NC(I)C(=O)NC1Cl' in combinations
-
-    combinations = cocktail.shake(compound_filters=["Drug-like"])
-    assert len(combinations) == 0
+    #
+    # cocktail = Cocktail(peptide_backbone, ligand_library = ["Br", "Cl", "I", "F"])
+    #
+    # combinations = cocktail.shake(compound_filters=["Lipinski"])
+    # assert 'O=C1NC(Br)C(=O)NC(I)C(=O)NC(Cl)C(=O)NC1F' in combinations
+    #
+    # combinations = cocktail.shake(compound_filters=["Ghose"])
+    # assert len(combinations) == 0
+    #
+    # combinations = cocktail.shake(compound_filters=["Veber"])
+    # assert 'O=C1NC(I)C(=O)NC(Br)C(=O)NC1F' in combinations
+    #
+    # combinations = cocktail.shake(compound_filters=["Rule of 3"])
+    # assert len(combinations) == 0
+    #
+    # combinations = cocktail.shake(compound_filters=["REOS"])
+    # assert 'O=C1NC(Br)C(=O)NC(I)C(=O)NC1Cl' in combinations
+    #
+    # combinations = cocktail.shake(compound_filters=["Drug-like"])
+    # assert len(combinations) == 0
