+++ requirements.txt
@@ -6,7 +6,7 @@ chardet==3.0.4
 idna==2.8
 imagesize==1.1.0
 importlib-metadata==0.23
-Jinja2==2.10.3
+Jinja2==2.11.1
 lxml==4.4.1
 MarkupSafe==1.1.1
 MolVS==0.1.1
