### 2020-02-09 #1992 @csoneson
👋 @greglandrum - would you be interested in reviewing this submission for the Journal of Open Source Software (JOSS)?

> Cocktail Shaker: An open source drug expansion and enumeration library for peptides

Cocktail Shaker is a python package for exploring, expanding, and synthesizing chemical peptide data, operating within the RDKit platform.

Software: https://github.com/Sulstice/cocktail-shaker
Short accompanying paper: https://github.com/openjournals/joss-papers/blob/joss.01992/joss.01992/10.21105.joss.01992.pdf

JOSS reviews are performed openly on GitHub and are based on a checklist that can be found here: https://joss.readthedocs.io/en/latest/review_checklist.html
Additional information about reviewing for JOSS is available via https://joss.readthedocs.io/en/latest/reviewer_guidelines.html
