### 2020-02-09 6789679
Update sphinx from 2.2.1 to 2.4.0
