### 2020-02-12 #1992 @csoneson
@greglandrum Please directly fill the checklist in the first post above so that we have a summary in one place.
You can then comment on individual items/aspects in separate comments (or open issues in the software repository and just mention them here).
