### 2020-02-16 #1992 @danielskatz
Yes, you should - the action @whedon took was, in the process of adding a reviewer, to unassign you than to reassign you.
