+++ requirements.txt
@@ -12,11 +12,8 @@ MarkupSafe==1.1.1
 MolVS==0.1.1
 more-itertools==7.2.0
 numpy==1.17.3
-olefile==0.46
 packaging==19.2
 pandas==0.25.3
-Pillow==6.2.1
-pluggy==0.13.1
 progressbar==2.5
 progressbar2==3.47.0
 py==1.8.0
