### 2020-02-19 4475e5d
Update sphinx from 2.2.1 to 2.4.2
