+++ requirements.txt
@@ -21,7 +21,7 @@ Pygments==2.4.2
 pyparsing==2.4.5
 pytest==5.3.0
 python-dateutil==2.8.1
-python-utils==2.3.0
+python-utils==2.4.0
 pytz==2019.3
 requests==2.22.0
 ruamel.yaml==0.16.5
