+++ README.md
@@ -1,14 +1,8 @@
-Cocktail Shaker: Drug Expansion and Enumeration for Peptides!
-=============================================================
+Megladon: The Protoss Shark AI here to challenge!
+=================================================
 
-[![Build](https://travis-ci.org/Sulstice/Cocktail-Shaker.svg?branch=master)](https://travis-ci.org/Sulstice/Cocktail-Shaker)
-[![License: MPL 2.0](https://img.shields.io/badge/License-MPL%202.0-brightgreen.svg)](https://opensource.org/licenses/MPL-2.0)
-[![Coverage](https://coveralls.io/repos/github/Sulstice/Cocktail-Shaker/badge.svg?branch=master)](https://coveralls.io/github/Sulstice/Cocktail-Shaker?branch=master)
+[![License: MPL 2.0](https://img.shields.io/badge/License-MPL%202.0-brightgreen.svg)](https://opensource.org/licenses/MPL-2.0))
 ![Python](https://img.shields.io/badge/python-3.6-blue.svg)
-[![Gitter](https://badges.gitter.im/Cocktail-Shaker/community.svg)](https://gitter.im/Cocktail-Shaker/community?utm_source=badge&utm_medium=badge&utm_campaign=pr-badge&utm_content=badge)
-[![Zenodo](https://zenodo.org/badge/170644606.svg)](https://zenodo.org/badge/latestdoi/170644606)
-[![Documentation Status](https://readthedocs.org/projects/cocktail-shaker/badge/?version=latest)](https://cocktail-shaker.readthedocs.io/en/latest/?badge=latest)
-[![status](https://joss.theoj.org/papers/c2e1d3c408a5729d832b34ac680d6305/status.svg)](https://joss.theoj.org/papers/c2e1d3c408a5729d832b34ac680d6305)
 
 <p align="center">
   <img width="200" height="300" src="images/logoshaker.png">
@@ -21,99 +15,44 @@ library**. cocktail-shaker leverages the computational power of
 Announcements
 =============
 
--   **Release!** Version 1.0.0-beta, August 26, 2019
--   **RDkit UGM 2019**: talk at September 25th at the University of
-    Hamburg, Germany.
+-   Work has began! March 2020
 
-Using Cocktail Shaker
+Using Megladon
 =====================
 
-cocktail-shaker is a young library under heavy development at this time.
-It targets two categories of users:
+Megladon will be my starcraft 2 ai that will be publically avaliable for people to use freely in building their AI. As the project evolves and the megladon gains more power and success certain features of the bot will be non-disclosed (I expect it to go pro - high expectations for my bot). 
 
-1.  **Users familiar with RDKit**, or those willing to learn RDKit, who want to
-    create fast sets of data for high throughput screening or machine
-    learning.
-2.  **Open-Science Scientists without any knowledge of RDKit** who are
-    seeking a a high-level wrapper to create chemical files for their
-    software.
+Megladon's goals:
 
-If you're in the first category, then you can already start using RDKit.
-cocktail-shaker offers a Pythonic, easy-to-use library and you can start
-channeling molecules in the expansion library. Instead of validating the
-sanity of the data, cocktail-shaker takes care of that for you. With
-each molecule being generated it will head into a 1D and/or 2D
-validation check (3D not supported yet).
-
-If you're in the second category, we're starting to build experimental
-high-level python code to take care a lot of the underpinnings of RDKit.
+- Become a grandmaster of the protoss race
 
 Installation 
-==================
+============
 
-Cocktail-shaker runs on Python 3.5+ and within a conda environment due to the RDKit dependency. A conda installation is coming soon (as soon as I figure it out). 
+Megladon will be running primarily with python 3.5+. As for now if you want to submit any PRs then please clone/fork this repository and follow the contribution guidelines. 
 
[…]
