### 2020-03-04 #1992 @Sulstice
@csoneson I updated per some of the requests from @greglandrum so there should be updates to some of the initial README and installation instructions. Please let me know if it isn't clear. For making the package a conda installable it will take some time (as I figure out how).
