### 2020-03-08 499430b
Update pygments from 2.4.2 to 2.6.1
