+++ requirements.txt
@@ -17,7 +17,7 @@ pandas==0.25.3
 progressbar==2.5
 progressbar2==3.47.0
 py==1.8.0
-Pygments==2.4.2
+Pygments==2.6.1
 pyparsing==2.4.5
 pytest==5.3.0
 python-dateutil==2.8.1
