+++ requirements.txt
@@ -13,7 +13,7 @@ MolVS==0.1.1
 more-itertools==7.2.0
 numpy==1.17.3
 packaging==19.2
-pandas==0.25.3
+pandas==1.0.2
 progressbar==2.5
 progressbar2==3.47.0
 py==1.8.0
