### 2020-03-24 b01e244
ss/MASTER fixed the documents and made it a little more clear on the quick start
