+++ docs/source/guide/quickstart.rst
@@ -9,12 +9,16 @@ This page instructs you on how to get started with cocktail-shaker. To begin, ma
 Basic Usage
 -----------
 
-The simplest way to use cocktail-shaker is to create a cocktail object with the ``shake`` function and create new compounds:
+The simplest way to use cocktail-shaker is to create a Peptide Builder object to generate your "base" peptide string,
+then to create cocktail object passing the peptide backbone and with the ``shake`` function you can generate your combinations:
 
     >>> from cocktail_shaker import Cocktail
     >>> from cocktail_shaker import PeptideBuilder
     >>> peptide_backbone = PeptideBuilder(2)
-    >>> cocktail = Cocktail(peptide_backbone,ligand_library = ['Br', 'I'])
+    >>> cocktail = Cocktail(
+    >>>                     peptide_backbone,
+    >>>                     ligand_library = ['Br', 'I']
+    >>>                     )
     >>> combinations = cocktail.shake()
     >>> print (combinations)
     >>> ['NC(Br)C(=O)NC(I)C(=O)NCC(=O)O', 'NC(I)C(=O)NC(Br)C(=O)NCC(=O)O']
@@ -24,3 +28,53 @@ Write the new compounds into an SDF file:
     >>> FileWriter('new_compounds', new_compounds, 'sdf')
 
 In this example, we have taken one SMILES string and expanded the compounds into a variety of variations into one SDF file.
+
+Cocktail Shaker uses the PeptideBuilder class to generate base peptide backbone strings that can be then passed as into the
+cocktail shaker object. Alternatively, a user can generate their own peptide string as long as it conforms with the cocktail shaker
+requirements. See the peptide builder documentation for more information.
+
+More Examples Peptide Builder
+-----------------------------
+
+Generation of a circular peptide
+
+>>> from cocktail_shaker import PeptideBuilder
+>>> peptide_molecule = PeptideBuilder(3, circular=True)
+>>> print (peptide_molecule)
+>>> O=C1C([*:1])NC(=O)C([*:2])NC(=O)C([*:3])N1
+
+Using the stereoisomer function with Cocktail Shaker
+
+>>> from cocktail_shaker import Cocktail
+>>> from cocktail_shaker import PeptideBuilder
+>>> peptide_backbone = PeptideBuilder(1)
+>>> cocktail = Cocktail(
+>>>     peptide_backbone,
+>>>     ligand_library = ['Br'],
+>>>     enable_isomers = True
+>>> )
+>>> combinations = cocktail.shake()
+>>> print (combinations)
+>>> ['N[C@H](Br)C(=O)NCC(=O)O', 'N[C@@H](Br)C(=O)NCC(=O)O']
+
+
+More Examples of Cocktail Shaker
+-----------
+
+Using the include amino acid function
+
+>>> from cocktail_shaker import Cocktail
+>>> from cocktail_shaker import PeptideBuilder
+>>> peptide_backbone = PeptideBuilder(1)
+>>> cocktail = Cocktail(
+>>>     peptide_backbone,
+>>>     include_amino_acids = True
+>>> )
+>>> combinations = cocktail.shake()
[…]
