### 2020-03-24 fe3177c
ss/MASTER fixed a class name in the docs for an example
