### 2020-03-30 74e2723
ss/MASTER added the docs an example that I am using it for
