+++ requirements.txt
@@ -18,7 +18,7 @@ progressbar==2.5
 progressbar2==3.47.0
 py==1.8.0
 Pygments==2.4.2
-pyparsing==2.4.5
+pyparsing==2.4.7
 pytest==5.3.0
 python-dateutil==2.8.1
 python-utils==2.3.0
