### 2020-04-13 #1992 @olivertomic
@arfon I am in touch with @Sulstice regarding examples of how to run the software. Work in progress. :-)
