### 2020-04-14 9adf984
ss/MASTER slight modifications and removed the test.py kicking off the pytest
