+++ requirements.txt
@@ -20,7 +20,7 @@ pyparsing==2.4.5
 pytest==5.3.0
 python-dateutil==2.8.1
 python-utils==2.3.0
-pytz==2019.3
+pytz==2020.1
 requests==2.22.0
 ruamel.yaml==0.16.5
 ruamel.yaml.clib==0.2.0
