### 2020-05-21 #1992 @Sulstice
Yes! I am making a little more interactive examples of how to use Cocktail Shaker per as @olivertomic suggestions (i.e juptyer notebooks or a gitpod) I am aiming for this saturday may 23rd (free time) to wrap that up.

I believe I tackled nearly all the issues attached with the repo filed from both greg and oliver and should be good to go unless I missed something (please let me know!).
