### 2020-05-23 #1992 @Sulstice
I added jupyter notebook examples for each of the apis with some more documentation into how it works with examples. I also added a litle ascii cinema in the README file for a quick visibility so someone can see if they would like to pip install it prior.

If there is anything else let me know!
