### 2020-06-02 #1992 @olivertomic
I am happy to approve the paper for publication. @Sulstice , and @csoneson . Just a minor change would be requried in paper.md itself.

The text says: ´Cocktail Shaker´ consists of **four** major class objects available to the user: PeptideMolecule, CocktailShaker, and FileWriter.´

However, only three are mentioned.

@Sulstice , thank you for the collaboration and implementation of suggestions.
