### 2020-06-02 #1992 @Sulstice
Yes you are definitely right, kind of an artifact of time of submission and me updating the modules. I will fix that later today!
