+++ requirements.txt
@@ -29,6 +29,6 @@ snowballstemmer==2.0.0
 Sphinx==2.2.1
 sphinx-bootstrap-theme==0.7.1
 urllib3==1.25.7
-wcwidth==0.1.7
+wcwidth==0.2.4
 zipp==0.6.0
 pyOpenSSL==19.1.0
