+++ requirements.txt
@@ -14,7 +14,7 @@ more-itertools==7.2.0
 numpy==1.17.3
 packaging==19.2
 pandas==0.25.3
-py==1.8.0
+py==1.8.2
 Pygments==2.4.2
 pyparsing==2.4.5
 pytest==5.3.0
