+++ requirements.txt
@@ -21,7 +21,7 @@ pytest==5.3.0
 python-dateutil==2.8.1
 python-utils==2.3.0
 pytz==2019.3
-requests==2.22.0
+requests==2.24.0
 ruamel.yaml==0.16.5
 ruamel.yaml.clib==0.2.0
 six==1.13.0
