### 2020-07-05 77f9488
ss/167 removed the proline check as it is unnecessary
