### 2020-07-05 e67a7bd
ss/167 fixed the paper.md file
