+++ paper.md
@@ -35,7 +35,7 @@ The toolkit offers a variety of utilities that includes: parsing and producing r
 designed for any chemical software, employing click chemistry methods for ease of exchange compounds, chemical data writing, 
 and chemical representation enumeration employed for machine learning.
 
-```Cocktail Shaker``` consists of four major class objects available to the user: PeptideMolecule, CocktailShaker, and FileWriter.
+```Cocktail Shaker``` consists of three major class objects available to the user: PeptideMolecule, CocktailShaker, and FileWriter.
 Using string manipulation Peptide Molecule can build SMILES strings with allocated slots defined by the user. The user can
 then enter the produced SMILES into the CocktailShaker object with a library of ligands represented by smiles and optional arguments
 of whether to include generation of stereoisomers and/or natural amino acids.  
