### 2020-07-05 #1992 @Sulstice
@richardjgowers Sorry! I don't know what happened to that fix I had (maybe forgot to commit). I resolved your issue as well - passed tests and released a 1.1.6 release. Let me know if there is anything else!
