+++ requirements.txt
@@ -7,7 +7,7 @@ idna==2.8
 imagesize==1.1.0
 importlib-metadata==0.23
 Jinja2==2.10.3
-lxml==4.4.1
+lxml==4.5.2
 MarkupSafe==1.1.1
 MolVS==0.1.1
 more-itertools==7.2.0
