### 2020-07-11 #1992 @richardjgowers
@Sulstice minor point, RDKit is written in C++ not C.  The paper references 3D generation using webchem, I can't find any references to this in the documentation?
