### 2020-07-11 #1992 @Sulstice
@richardjgowers Good point, I'll fix immediately.

Huh, I actually didn't mention how the FileWriter module works. It's another cirpy (https://github.com/mcs07/CIRpy) under the hood (but more robust) hitting the NIH cactus resolver. I'm going to add that to the docs for how the FileWriter works - does that suffice?
