+++ docs/source/guide/cocktail.rst
@@ -6,29 +6,10 @@ Cocktail API
 This page introduces the functionality of the cocktail object and provides a deeper look into what it can accomplish in the future.
 
 
-The Cocktail Class
-------------------
+The GlobalChem Class
+--------------------
 
-The cocktail object is the heart of the package and allows you to parse in a peptide SMILES string and expand your drug
-library with different functional groups as well as enumerate representations of the molecule in 1D and 2D.
-
-You instantiate a ``Cocktail`` object by parsing in a list of SMILES.
-Cocktail-Shaker will already handle the SMILES to RDKit Mol Object for you without having to subject to create them
-yourself.
-If the SMILES fail to load and is *not* supported, then ``MoleculeError`` will be raised instead.
-
->>> from cocktail_shaker import Cocktail
->>> from cocktail_shaker import PeptideBuilder
->>> peptide_molecule = PeptideBuilder(2)
->>> cocktail = Cocktail(peptide_molecule, ligand_library=['Br', 'I'])
->>> print (cocktail)
->>> CocktailObject
-
-cocktail-shaker, under the hood, uses two validation methods to determine whether a molecule is a legitimate molecule.
-One method uses the MolVS validator and the other using RDKit's rendering capabilities from SMILES to RDKit Mol Object, we can determine
-if a generated molecule is legitimate. By using both validation methods, cocktail-shaker can ensure validation of the SMILES.
-
-.. attribute:: peptide_molecule
+.. attribute:: amino_acid_side_chains
 
   The peptide backbone string generated from PeptideBuilder
 
@@ -44,156 +25,3 @@ if a generated molecule is legitimate. By using both validation methods, cocktai
 
   Include the natural amino acids except for Proline (TBD), list of smiles found in :ref:`amino acids <aminoacids>`.
   Defaults to False.
-
-
-The "Shake" Function
---------------------
-
-The shake module detects functional groups present on a molecule, breaks their bond, and then adds a functional
-group (not itself) from the datasource library as replacement.
-
-Essentially, the shake compound is utilizing a chemical method known as "click" chemistry in the snapping and formation
-of new bonds linking them together.
-
-You instantiate a ``Cocktail`` object by parsing in a list of SMILES and then "shake" the compounds.
-cocktail-shaker shows the functional groups that can be detected and then swaps them accordingly.
-If the shake fails to work, then ``MoleculeError`` will be raised instead. In this case, contact the lead developer.
-
->>> from cocktail_shaker import Cocktail
->>> from cocktail_shaker import PeptideBuilder
->>> peptide_molecule = PeptideBuilder(2)
->>> cocktail = Cocktail(peptide_molecule, ligand_library=['Br', 'I'])
->>> print (cocktail.shake())
->>> ['NC(Br)C(=O)NC(I)C(=O)NCC(=O)O', 'NC(I)C(=O)NC(Br)C(=O)NCC(=O)O']
-
-As mentioned before, validation occurs internally. This means that the molecules being generated are validated
-before appended the list.
-
-There are some minor restrictions in the first release of this package as described below.
[…]
