### 2020-07-11 #1992 @Sulstice
I added it in the docs a little bit here: https://cocktail-shaker.readthedocs.io/en/latest/guide/file_handling.html#the-filewriter-module
