+++ requirements.txt
@@ -28,7 +28,7 @@ six==1.13.0
 snowballstemmer==2.0.0
 Sphinx==2.2.1
 sphinx-bootstrap-theme==0.7.1
-urllib3==1.25.7
+urllib3==1.25.10
 wcwidth==0.1.7
 zipp==0.6.0
 pyOpenSSL==19.1.0
