### 2020-07-25 #1992 @richardjgowers
@Sulstice the [mol2 parser](https://github.com/Sulstice/cocktail-shaker/blob/master/cocktail_shaker/file_handler.py#L290) seems like it might not work as advertised in the docs.  Similarly [there's two method here](https://github.com/Sulstice/cocktail-shaker/blob/master/cocktail_shaker/peptide_builder.py#L161) that look like they're yet to be implemented.  If they're not done yet put up a `NotImplementedError` or just remove them.
