### 2020-07-26 9467b71
ss/patches - not implemented error, new docs
