+++ docs/source/guide/file_formats.rst
@@ -12,8 +12,6 @@ Input can be parsed in a variety of file formats that are autodetected::
 The full list of file formats that cocktail-shaker supports parsing::
 
     sdf       # 2D Structure file format
-    mol2      # (Coming soon)
-
 
 Output can be returned in a variety of file formats that are specified your enumerated compounds and your
 desired file extension::
