+++ requirements.txt
@@ -13,7 +13,7 @@ MolVS==0.1.1
 more-itertools==7.2.0
 numpy==1.17.3
 packaging==19.2
-pandas==0.25.3
+pandas==1.1.0
 py==1.8.0
 Pygments==2.4.2
 pyparsing==2.4.5
