+++ requirements.txt
@@ -17,7 +17,7 @@ pandas==0.25.3
 py==1.8.0
 Pygments==2.4.2
 pyparsing==2.4.5
-pytest==5.3.0
+pytest==6.0.1
 python-dateutil==2.8.1
 python-utils==2.3.0
 pytz==2019.3
