### 2020-08-09 8e86833
Small changes to paper and bib file
