+++ paper.bib
@@ -1,21 +1,27 @@
 @article{Craig:2013-1,
-    author = {{Craig}, D. and {Fairlie}, D. and {Liras} S. and {Price} D.},
+    author = {{Craig}, D.J. and {Fairlie}, D.P. and {Liras}, S. and {Price}, D.},
     title = "The future of peptide-based drugs.",
     journal = "Chemical Biology & Drug Design",
+    volume = 81,
+    issue = 1,
+    pages = {136--147},
     year = "2013",
     doi = "10.1111/cbdd.12055"
 }
 
 @article{Bruno:2013-2,
-    author = {{Bruno}, B. and {Miller}, G. and {Lim} C.},
+    author = {{Bruno}, B.J. and {Miller}, G.D. and {Lim}, C.S.},
     title = "Basics and recent advances in peptide and protein drug delivery",
     journal = "Therapeutic Delivery",
+    volume = 4,
+    issue = 11,
+    pages = {1443--1467},
     year = "2013",
     doi = "10.4155/tde.13.104"
 }
 
 @article{Reynolds:2010-3,
-    author = {{Reynolds}, C. and {Merz}, K. and {Ringe} D.},
+    author = {{Reynolds}, C. and {Merz}, K. and {Ringe}, D.},
     title = "Drug Design: Structure- and Ligand-Based Approaches",
     journal = "Cambrige University Press",
     year = "2010",
@@ -24,9 +30,12 @@
 
 @article{Maillet:2020-4,
     author = {{Malliet}, N.},
-    title = "Generator: fast and efficient in silico protein digestion",
+    title = "Rapid Peptides Generator: fast and efficient in silico protein digestion",
     journal = "NAR Genomics and Bioinformatics",
     year = "2020",
+    volume = 2,
+    issue = 1,
+    note = {lqz004},
     doi = "10.1093/nargab/lqz004"
 }
 
