### 2020-08-09 190dbd4
A couple of more small edits
