### 2020-08-09 3c6290a
Small fixes to bib file
