+++ paper.bib
@@ -23,14 +23,14 @@
 @article{Reynolds:2010-3,
     author = {{Reynolds}, C. and {Merz}, K. and {Ringe}, D.},
     title = "Drug Design: Structure- and Ligand-Based Approaches",
-    journal = "Cambrige University Press",
+    journal = "Cambridge University Press",
     year = "2010",
     ISBN = "978-0521887236"
 }
 
 @article{Maillet:2020-4,
     author = {{Malliet}, N.},
-    title = "Rapid Peptides Generator: fast and efficient in silico protein digestion",
+    title = "{Rapid Peptides Generator}: fast and efficient in silico protein digestion",
     journal = "NAR Genomics and Bioinformatics",
     year = "2020",
     volume = 2,
