### 2020-08-09 #1992 @csoneson
👋 @Sulstice - I have made a pull request to your repository with a few small fixes to the paper and the bib file. Could you please check that they look ok, that I didn't misinterpret anything, and merge it. Please also clarify this sentence in the introduction, I'm not quite sure how to interpret it:

>  With increasing interest eludes to increasing high throughput screening of peptides computationally.
