### 2020-08-10 #1992 @csoneson
@Sulstice - the paper looks good! Here are the next steps:

- [ ] Create a tagged release of the software, and report back the version tag here
- [ ] Archive the release in Zenodo
- [ ] Check that the Zenodo archive has the same title and author list as the paper.
- [ ] Report back the Zenodo DOI here
