### 2020-08-13 #1992 @csoneson
Thanks. Could you please fix the metadata of the Zenodo archive so that it has the same title and author list as the JOSS paper.
Also, the paper says that "it will now be released with version 1.1.0", while the current version is 1.1.8 - I would suggest changing the formulation in the paper to avoid confusion.
