### 2020-08-17 #1992 @kyleniemeyer
Hi @Sulstice, I'm doing some final review steps—can you add more info to the RDKit citation? It looks like you can add either a DOI for the most recent version (https://zenodo.org/record/3981263), or for all versions (10.5281/zenodo.591637)
