+++ paper.bib
@@ -43,5 +43,6 @@
     author = {{Landrum}, G.},
     title = "RDKit: Open-Source Cheminformatics Software",
     year = "2019",
+    doi = "10.5281/zenodo.591637"
 }
 
