### 2020-08-17 #1992 @Sulstice
@kyleniemeyer Done :) I added the DOI for all versions and in the docs a `RDKit 2019.09.1+` -> ran tests on latest RDKit 2020 release and passing.
